Refactoring : move command building to a new source file
